response 구조 변경 및 service에서 dto로 받던거 vo로 받게 리팩토링

+++ src/main/java/com/hh/board/domain/post/PostController.java
@@ -3,20 +3,29 @@ package com.hh.board.domain.post;
 
 import com.hh.board.common.dto.SearchDto;
 import com.hh.board.common.file.FileUtils;
+import com.hh.board.common.paging.Pagination;
 import com.hh.board.common.paging.PagingResponse;
 import com.hh.board.common.response.Response;
+import com.hh.board.common.vo.SearchVo;
 import com.hh.board.domain.file.FileRequestDto;
 import com.hh.board.domain.file.FileResponseDto;
 import com.hh.board.domain.file.FileService;
+import com.hh.board.domain.file.FileVo;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
+import static com.hh.board.common.vo.SearchVo.toVo;
+import static com.hh.board.domain.file.FileVo.toVo;
 import static com.hh.board.domain.post.PostVo.*;
 
 @CrossOrigin(origins = "*")
@@ -35,11 +44,23 @@ public class PostController {
     @GetMapping("/posts")
     public ResponseEntity<Response> getPosts(SearchDto searchDto) {
 
-        PagingResponse<PostResponseDto> listWithPagination = postService.findAllPostBySearch(searchDto);
+        // 빈 반환 객체 생성
+        // 조건에 해당하는 데이터가 없는 경우, 응답 데이터에 비어있는 객체를 담아 반환
+        PagingResponse<PostResponseDto> listWithPagination = new PagingResponse<>(Collections.emptyList(), null);
+
+        int count = postService.countAllPostBySearch(toVo(searchDto));
+
+        if(count >= 1) {
+            // Pagination 객체를 생성해서 페이지 정보 계산 후 SearchDto 타입 객체에 계산된 페이지 정보 저장
+            // pagination 객체 생성자에 searchDto.setPagination 포함
+            Pagination pagination = new Pagination(count, searchDto);
+
+            listWithPagination = postService.findAllPostBySearch(toVo(searchDto), pagination);
+        }
 
         return ResponseEntity
                 .status(HttpStatus.OK)
-                .body(Response.success(listWithPagination));
+                .body(new Response(listWithPagination));
     }
 
     // 게시글 조회
@@ -52,20 +73,31 @@ public class PostController {
 
         return ResponseEntity
                 .status(HttpStatus.OK)
-                .body(Response.success(postResponseDto));
+                .body(new Response(postResponseDto));
     }
 
     // 게시글 저장
     @PostMapping("/posts")
     public ResponseEntity<Response> savePost(@Valid PostRequestDto postRequestDto) {
 
-        int postId = postService.savePost(toVo(postRequestDto));
-        List<FileRequestDto> files = fileUtils.uploadFiles(postRequestDto.getFiles());
-        fileService.saveFiles(postId, files);
+        int postId = postService.savePost(PostVo.toVo(postRequestDto));
+
+        List<MultipartFile> files = postRequestDto.getFiles();
+        if(!CollectionUtils.isEmpty(files)) {
+            List<FileRequestDto> uploadFiles = fileUtils.uploadFiles(files);
+
+            List<FileVo> fileVoList = new ArrayList<>();
+            uploadFiles.stream().forEach(f -> {
+                f.setPostId(postId);
+                fileVoList.add(toVo(f));
+            });
+
+            fileService.saveFiles(fileVoList);
+        }
 
         return ResponseEntity
                 .status(HttpStatus.CREATED)
-                .body(Response.success(postId));
+                .body(new Response(postId + "번 게시글 생성 완료"));
     }
 
     // 게시글 수정
@@ -73,21 +105,36 @@ public class PostController {
     public ResponseEntity<Response> updatePost(@PathVariable int postId, @Valid PostRequestDto postRequestDto) {
 
         // 1. 게시글 정보 수정
-        postService.updatePost(toVo(postRequestDto));
-        // 2. 파일 업로드 (to disk)
-        List<FileRequestDto> uploadFiles = fileUtils.uploadFiles(postRequestDto.getFiles());
-        // 3. 파일 정보 저장 (to database)
-        fileService.saveFiles(postRequestDto.getPostId(), uploadFiles);
-        // 4. 삭제할 파일 정보 조회 (from database)
-        List<FileResponseDto> deleteFiles = fileService.findAllFileByIds(postRequestDto.getRemoveFileIds());
-        // 5. 파일 삭제 (from disk)
-        fileUtils.deleteFiles(deleteFiles);
-        // 6. 파일 삭제 (from database)
-        fileService.deleteAllFileByIds(postRequestDto.getRemoveFileIds());
+        postService.updatePost(PostVo.toVo(postRequestDto));
+
+        List<MultipartFile> files = postRequestDto.getFiles();
+        if(!CollectionUtils.isEmpty(files)) {
+            // 2. 파일 업로드 (to disk)
+            List<FileRequestDto> uploadFiles = fileUtils.uploadFiles(files);
+
+            List<FileVo> fileVoList = new ArrayList<>();
+            uploadFiles.stream().forEach(f -> {
+                f.setPostId(postId);
+                fileVoList.add(toVo(f));
+            });
+
+            // 3. 파일 정보 저장 (to database)
+            fileService.saveFiles(fileVoList);
+        }
+
+        List<Integer> removeFileIds = postRequestDto.getRemoveFileIds();
+        if(!CollectionUtils.isEmpty(removeFileIds)) {
+            // 4. 삭제할 파일 정보 조회 (from database)
+            List<FileResponseDto> deleteFiles = fileService.findAllFileByIds(removeFileIds);
+            // 5. 파일 삭제 (from disk)
+            fileUtils.deleteFiles(deleteFiles);
+            // 6. 파일 삭제 (from database)
+            fileService.deleteAllFileByIds(removeFileIds);
+        }
 
         return ResponseEntity
                 .status(HttpStatus.OK)
-                .body(Response.success(postId + "번 게시글 수정 완료"));
+                .body(new Response(postId + "번 게시글 수정 완료"));
     }
 
     // 게시글 삭제
@@ -98,7 +145,7 @@ public class PostController {
 
         return ResponseEntity
                 .status(HttpStatus.OK)
-                .body(Response.success(postId + "번 게시글 삭제 완료"));
+                .body(new Response(postId + "번 게시글 삭제 완료"));
     }
